+++ package.json
@@ -17,6 +17,7 @@
     "spellcheck": "cspell \"**/*.{tsx,ts,md}\" --show-suggestions --no-progress",
     "test": "jest -i --no-cache",
     "test:coverage": "jest -i --no-cache --coverage",
+    "test:e2e": "pnpm test:core:e2e && pnpm test:widgets:e2e",
     "test:core": "jest -i --no-cache --selectProjects \"UI Kit Core\"",
     "test:utils": "jest -i --no-cache --selectProjects \"UI Kit Utils\"",
     "test:hooks": "jest -i --no-cache --selectProjects \"UI Kit Hooks\"",
