+++ packages/core/src/components/Table/Table.spec.tsx
@@ -1,13 +1,12 @@
-import { screen } from '../../../customTest';
-
+import { act } from 'react-dom/test-utils';
 import Table from '@components/Table';
 import TableHead from '@components/TableHead';
 import TableRow from '@components/TableRow';
 import TableCell from '@components/TableCell';
 import TableBody from '@components/TableBody';
-import { TableSortingComponent } from './components/TableSortingComponent';
 import { fireEvent } from '@storybook/testing-library';
-import { act } from 'react-dom/test-utils';
+import { SortableTable } from './stories/SortableTable';
+import { screen } from '../../../customTest';
 
 describe('Table', () => {
   it('Render table', () => {
@@ -42,44 +41,46 @@ describe('Table', () => {
     expect(cells.length).toEqual(9);
   });
 
-  it('Render Sorting table', () => {
-    render(<TableSortingComponent />);
+  describe('Sorting Table', () => {
+    it('Render table', () => {
+      render(<SortableTable />);
 
-    const cells = screen.queryAllByRole('cell');
+      const cells = screen.queryAllByRole('cell');
 
-    expect(screen.getByRole('table')).toBeInTheDocument();
-    expect(screen.getByTitle('Arrow Down')).toBeInTheDocument();
-    expect(cells.length).toEqual(33);
-  });
-
-  it('Should sort changing worked', () => {
-    const callback = jest.fn();
-    render(<TableSortingComponent onSortingChange={callback} />);
-
-    const cells = screen.queryAllByRole('cell');
-
-    act(() => {
-      fireEvent.click(cells[0]);
-    });
-    expect(callback).toHaveBeenLastCalledWith({
-      column: 'title',
-      order: 'asc',
+      expect(screen.getByRole('table')).toBeInTheDocument();
+      expect(screen.getByTitle('Arrow Down')).toBeInTheDocument();
+      expect(cells.length).toEqual(33);
     });
 
-    act(() => {
-      fireEvent.click(cells[0]);
-    });
-    expect(callback).toHaveBeenLastCalledWith({
-      column: 'title',
-      order: 'desc',
-    });
+    it('Should change sorting correctly', () => {
+      const callback = jest.fn();
+      render(<SortableTable onSortingChange={callback} />);
 
-    act(() => {
-      fireEvent.click(cells[0]);
-    });
-    expect(callback).toHaveBeenLastCalledWith({
-      column: 'title',
-      order: 'asc',
+      const cells = screen.queryAllByRole('cell');
+
+      act(() => {
+        fireEvent.click(cells[0]);
+      });
+      expect(callback).toHaveBeenLastCalledWith({
+        column: 'title',
+        order: 'asc',
+      });
+
+      act(() => {
+        fireEvent.click(cells[0]);
+      });
+      expect(callback).toHaveBeenLastCalledWith({
+        column: 'title',
+        order: 'desc',
+      });
+
+      act(() => {
+        fireEvent.click(cells[0]);
+      });
+      expect(callback).toHaveBeenLastCalledWith({
+        column: 'title',
+        order: 'asc',
+      });
     });
   });
 });
