+++ packages/core/src/components/Table/stories/consts.ts
@@ -1,0 +1,6 @@
+import { SortInfo } from './types';
+
+export const DEFAULT_SORT: SortInfo = {
+  column: 'value',
+  order: 'desc',
+};
