+++ packages/core/src/components/Table/stories/types.ts
@@ -1,0 +1,9 @@
+import { RowKeys } from './mock';
+
+export type RowKeysArray = ['title', 'value', 'description'];
+export type SortInfo =
+  | {
+      column: RowKeys;
+      order: 'asc' | 'desc';
+    }
+  | Record<string, never>;
